requirements-docker-cuda11.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
biopython==1.81
blinker==1.4
boto3==1.26.53
botocore==1.29.53
bravado==11.0.3
bravado-core==5.17.1
brotlipy==0.7.0
cachetools==4.2.2
captum==0.8.0
casanovo==3.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
comet-ml==3.31.22
comm==0.1.2
configobj==5.0.6
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
Deprecated==1.2.13
depthcharge-ms==0.1.0
dulwich==0.20.50
einops==0.6.0
entrypoints==0.4
everett==3.1.0
executing==0.8.3
fastobo==0.12.2
flit_core==3.6.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.6.0
google-auth-oauthlib==0.4.4
idna==3.4
ipykernel==6.19.2
ipython==8.7.0
isoduration==20.11.0
jedi==0.18.1
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonref==1.1.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.1
kiwisolver==1.4.4
lark==1.1.5
lightning-utilities==0.5.0
Markdown==3.4.1
MarkupSafe==2.1.1
matchms==0.18.0
monotonic==1.6
msgpack==1.0.4
multidict==6.0.4
neptune-client==0.16.16
nest-asyncio==1.5.6
networkx==3.1
numba
numpy
oauthlib==3.2.1
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pickydict==0.4.0
platformdirs==2.5.2
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.57
Pygments==2.11.2
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pyteomics==4.5.6
python-dateutil==2.8.2
pytorch-lightning==1.8.6
pytorch-ranger==0.1.1
pytz==2022.7
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
semantic-version==2.10.0
sentry-sdk==1.12.1
setuptools==65.5.0
simplejson==3.18.1
six==1.16.0
smmap==5.0.0
sparsestack==0.4.1
spectrum-utils==0.4.1
stack-data==0.2.0
swagger-spec-validator==3.0.3
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
torch-optimizer==0.3.0
torchmetrics==0.11.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.13
wcwidth==0.2.5
webcolors==1.12
websocket-client==1.3.3
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
wurlitzer==3.0.3
yapf==0.32.0
yarl==1.8.2
lmdb==1.7.3
cupy-cuda11x==13.6.0
pydantic==1.10.24
pyarrow==23.0.0
